add calendar event add

+++ src/api/inc/calendar_database.h
@@ -8,7 +8,7 @@ void calendar_database_deinit(void);
 
 void calendar_database_init(void);
 
-Bool calendar_data_base_event_add(uint32_t weekday, uint32_t start_time, uint32_t stop_time, char8_t *event_name);
+Bool calendar_data_base_event_add(uint32_t weekday, float32_t start_time, float32_t stop_time, char8_t *event_name);
 
 
 #endif
