+++ src/api/inc/user_input_parse.h
@@ -14,7 +14,7 @@ void user_input_thread_init(void);
 
 void parse_file(void);
 
-inline Bool is_valid_time_test_wrapper(char8_t *word, uint32_t *start_time, uint32_t *stop_time);
+inline Bool is_valid_time_test_wrapper(char8_t *word, float32_t *start_time, float32_t *stop_time);
 
 inline uint32_t is_weekday_test_wrapper(char *day);
